Fixed issue #62
As shown in issue #62, for some inputs, the function generate_output_dataframes() from aggregation.py was raising an error because values in column 'struct_num_col' were being cast as string, while the code expects integer values.

So to fix this issue, I added a code to force the casting of this column to integer at line #391.